03/requirements.txt
black
Django==4.2
flake8
nltk
pandas
scikit-learn
python-docx